Fix issues involving file downloads
The file is downloaded to a temporary file first, then moved to the actual, desired target. The move is performed inside of the context manager of the NamedTemporaryFile. This means that the move function, shutil.move, attempts to delete (via os.unlink) the temporary file while it is still kept open. This causes an error (at least on Windows).
The issue is fixed by performing the move *outside* of the context manager, so that there is nobody who is keeping the file open when shutil.move wants to os.unlink it.
Another minor issue was that os.stat does not return the correct file size while the temporary file is not closed yet. The file size has to be retrieved via tell.